initialization of core
