+++ hrosailing/core/__init__.py
@@ -1,0 +1,5 @@
+import hrosailing.core.modelfunctions
+import hrosailing.core.computing
+import hrosailing.core.constants
+import hrosailing.core.data
+import hrosailing.core.statistics
